requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
cached-property==1.5.2
certifi==2021.5.30
dataclasses==0.8
h5py==2.10.0
html5lib==0.9999999
Keras==2.0.7
Markdown==2.2.0
numpy==1.19.5
Pillow==5.4.0
protobuf==3.9.2
PyYAML==6.0.1
scipy==1.5.4
six==1.16.0
tensorflow==1.2.0
Theano==1.0.5
Werkzeug==2.0.3
